feat: Color 값의 위치를 Model 에서 View 단으로 변경
